adding search via middleman-search gem
this adds a pretty neat search feature using the middleman-search gem,
which in turn uses lunr.js, a lightweight solr like search engine for
javascript. It should be noted that search will not work when modifying
the site using the middleman server, it only creates the search engine
on middleman build.

Also cleaning up a few things as well, including adding a better deploy
process to sync everything to an s3 bucket.